Make test order consistent even when tests are added or removed.
Often when tests have an order problem you want to remove tests to see
which two (or more) tests are causing the problem. However previously
adding or removing any tests would cause the order to completely change.
This change keeps the relative order of the tests the same to ease this
debugging.